+++ coresimd/x86/abm.rs
@@ -26,15 +26,15 @@ use stdsimd_test::assert_instr;
 #[inline]
 #[target_feature(enable = "lzcnt")]
 #[cfg_attr(test, assert_instr(lzcnt))]
-pub unsafe fn _lzcnt_u64(x: u64) -> u64 {
-    x.leading_zeros() as u64
+pub unsafe fn _lzcnt_u32(x: u32) -> u32 {
+    x.leading_zeros()
 }
 
 /// Counts the bits that are set.
 #[inline]
 #[target_feature(enable = "popcnt")]
 #[cfg_attr(test, assert_instr(popcnt))]
-pub unsafe fn _popcnt64(x: i64) -> i32 {
+pub unsafe fn _popcnt32(x: i32) -> i32 {
     x.count_ones() as i32
 }
 
@@ -45,12 +45,12 @@ mod tests {
     use coresimd::x86::*;
 
     #[simd_test = "lzcnt"]
-    unsafe fn test_lzcnt_u64() {
-        assert_eq!(_lzcnt_u64(0b0101_1010), 57);
+    unsafe fn test_lzcnt_u32() {
+        assert_eq!(_lzcnt_u32(0b0101_1010), 25);
     }
 
     #[simd_test = "popcnt"]
-    unsafe fn test_popcnt64() {
-        assert_eq!(_popcnt64(0b0101_1010), 4);
+    unsafe fn test_popcnt32() {
+        assert_eq!(_popcnt32(0b0101_1010), 4);
     }
 }
