+++ coresimd/x86/tbm.rs
@@ -263,41 +263,41 @@ pub unsafe fn _tzmsk_u64(x: u64) -> u64 {
 mod tests {
     use stdsimd_test::simd_test;
 
-    use coresimd::x86::i586::tbm;
+    use coresimd::x86::*;
 
     /*
     #[simd_test = "tbm"]
-    unsafe fn _bextr_u32() {
-        assert_eq!(tbm::_bextr_u32(0b0101_0000u32, 4, 4), 0b0000_0101u32);
+    unsafe fn test_bextr_u32() {
+        assert_eq!(_bextr_u32(0b0101_0000u32, 4, 4), 0b0000_0101u32);
     }
 
     #[simd_test = "tbm"]
-    unsafe fn _bextr_u64() {
-        assert_eq!(tbm::_bextr_u64(0b0101_0000u64, 4, 4), 0b0000_0101u64);
+    unsafe fn test_bextr_u64() {
+        assert_eq!(_bextr_u64(0b0101_0000u64, 4, 4), 0b0000_0101u64);
     }
     */
 
     #[simd_test = "tbm"]
-    unsafe fn _blcfill_u32() {
-        assert_eq!(tbm::_blcfill_u32(0b0101_0111u32), 0b0101_0000u32);
-        assert_eq!(tbm::_blcfill_u32(0b1111_1111u32), 0u32);
+    unsafe fn test_blcfill_u32() {
+        assert_eq!(_blcfill_u32(0b0101_0111u32), 0b0101_0000u32);
+        assert_eq!(_blcfill_u32(0b1111_1111u32), 0u32);
     }
 
     #[simd_test = "tbm"]
     #[cfg(not(target_arch = "x86"))]
-    unsafe fn _blcfill_u64() {
-        assert_eq!(tbm::_blcfill_u64(0b0101_0111u64), 0b0101_0000u64);
-        assert_eq!(tbm::_blcfill_u64(0b1111_1111u64), 0u64);
+    unsafe fn test_blcfill_u64() {
+        assert_eq!(_blcfill_u64(0b0101_0111u64), 0b0101_0000u64);
+        assert_eq!(_blcfill_u64(0b1111_1111u64), 0u64);
     }
 
     #[simd_test = "tbm"]
-    unsafe fn _blci_u32() {
+    unsafe fn test_blci_u32() {
         assert_eq!(
-            tbm::_blci_u32(0b0101_0000u32),
+            _blci_u32(0b0101_0000u32),
             0b1111_1111_1111_1111_1111_1111_1111_1110u32
         );
         assert_eq!(
-            tbm::_blci_u32(0b1111_1111u32),
+            _blci_u32(0b1111_1111u32),
             0b1111_1111_1111_1111_1111_1110_1111_1111u32
         );
     }
@@ -305,61 +305,61 @@ mod tests {
     #[simd_test = "tbm"]
     #[cfg(not(target_arch = "x86"))]
     #[cfg_attr(rustfmt, rustfmt_skip)]
-    unsafe fn _blci_u64() {
+    unsafe fn test_blci_u64() {
         assert_eq!(
-            tbm::_blci_u64(0b0101_0000u64),
+            _blci_u64(0b0101_0000u64),
             0b1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1110u64
         );
         assert_eq!(
-            tbm::_blci_u64(0b1111_1111u64),
+            _blci_u64(0b1111_1111u64),
             0b1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1110_1111_1111u64
         );
     }
 
     #[simd_test = "tbm"]
-    unsafe fn _blcic_u32() {
-        assert_eq!(tbm::_blcic_u32(0b0101_0001u32), 0b0000_0010u32);
-        assert_eq!(tbm::_blcic_u32(0b1111_1111u32), 0b1_0000_0000u32);
+    unsafe fn test_blcic_u32() {
+        assert_eq!(_blcic_u32(0b0101_0001u32), 0b0000_0010u32);
+        assert_eq!(_blcic_u32(0b1111_1111u32), 0b1_0000_0000u32);
     }
 
     #[simd_test = "tbm"]
     #[cfg(not(target_arch = "x86"))]
-    unsafe fn _blcic_u64() {
-        assert_eq!(tbm::_blcic_u64(0b0101_0001u64), 0b0000_0010u64);
-        assert_eq!(tbm::_blcic_u64(0b1111_1111u64), 0b1_0000_0000u64);
+    unsafe fn test_blcic_u64() {
+        assert_eq!(_blcic_u64(0b0101_0001u64), 0b0000_0010u64);
+        assert_eq!(_blcic_u64(0b1111_1111u64), 0b1_0000_0000u64);
     }
 
     #[simd_test = "tbm"]
-    unsafe fn _blcmsk_u32() {
-        assert_eq!(tbm::_blcmsk_u32(0b0101_0001u32), 0b0000_0011u32);
-        assert_eq!(tbm::_blcmsk_u32(0b1111_1111u32), 0b1_1111_1111u32);
-    }
-
-    #[simd_test = "tbm"]
-    #[cfg(not(target_arch = "x86"))]
-    unsafe fn _blcmsk_u64() {
-        assert_eq!(tbm::_blcmsk_u64(0b0101_0001u64), 0b0000_0011u64);
-        assert_eq!(tbm::_blcmsk_u64(0b1111_1111u64), 0b1_1111_1111u64);
-    }
-
-    #[simd_test = "tbm"]
-    unsafe fn _blcs_u32() {
-        assert_eq!(tbm::_blcs_u32(0b0101_0001u32), 0b0101_0011u32);
-        assert_eq!(tbm::_blcs_u32(0b1111_1111u32), 0b1_1111_1111u32);
+    unsafe fn test_blcmsk_u32() {
+        assert_eq!(_blcmsk_u32(0b0101_0001u32), 0b0000_0011u32);
+        assert_eq!(_blcmsk_u32(0b1111_1111u32), 0b1_1111_1111u32);
     }
 
     #[simd_test = "tbm"]
     #[cfg(not(target_arch = "x86"))]
-    unsafe fn _blcs_u64() {
-        assert_eq!(tbm::_blcs_u64(0b0101_0001u64), 0b0101_0011u64);
-        assert_eq!(tbm::_blcs_u64(0b1111_1111u64), 0b1_1111_1111u64);
+    unsafe fn test_blcmsk_u64() {
+        assert_eq!(_blcmsk_u64(0b0101_0001u64), 0b0000_0011u64);
+        assert_eq!(_blcmsk_u64(0b1111_1111u64), 0b1_1111_1111u64);
     }
 
     #[simd_test = "tbm"]
-    unsafe fn _blsfill_u32() {
-        assert_eq!(tbm::_blsfill_u32(0b0101_0100u32), 0b0101_0111u32);
+    unsafe fn test_blcs_u32() {
+        assert_eq!(_blcs_u32(0b0101_0001u32), 0b0101_0011u32);
+        assert_eq!(_blcs_u32(0b1111_1111u32), 0b1_1111_1111u32);
+    }
+
+    #[simd_test = "tbm"]
+    #[cfg(not(target_arch = "x86"))]
+    unsafe fn test_blcs_u64() {
+        assert_eq!(_blcs_u64(0b0101_0001u64), 0b0101_0011u64);
+        assert_eq!(_blcs_u64(0b1111_1111u64), 0b1_1111_1111u64);
+    }
+
+    #[simd_test = "tbm"]
+    unsafe fn test_blsfill_u32() {
+        assert_eq!(_blsfill_u32(0b0101_0100u32), 0b0101_0111u32);
         assert_eq!(
-            tbm::_blsfill_u32(0u32),
+            _blsfill_u32(0u32),
             0b1111_1111_1111_1111_1111_1111_1111_1111u32
         );
     }
@@ -367,22 +367,22 @@ mod tests {
     #[simd_test = "tbm"]
     #[cfg(not(target_arch = "x86"))]
     #[cfg_attr(rustfmt, rustfmt_skip)]
-    unsafe fn _blsfill_u64() {
-        assert_eq!(tbm::_blsfill_u64(0b0101_0100u64), 0b0101_0111u64);
+    unsafe fn test_blsfill_u64() {
+        assert_eq!(_blsfill_u64(0b0101_0100u64), 0b0101_0111u64);
         assert_eq!(
-            tbm::_blsfill_u64(0u64),
+            _blsfill_u64(0u64),
             0b1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111u64
         );
     }
 
     #[simd_test = "tbm"]
-    unsafe fn _blsic_u32() {
+    unsafe fn test_blsic_u32() {
         assert_eq!(
-            tbm::_blsic_u32(0b0101_0100u32),
+            _blsic_u32(0b0101_0100u32),
             0b1111_1111_1111_1111_1111_1111_1111_1011u32
         );
         assert_eq!(
-            tbm::_blsic_u32(0u32),
+            _blsic_u32(0u32),
             0b1111_1111_1111_1111_1111_1111_1111_1111u32
         );
     }
@@ -390,25 +390,25 @@ mod tests {
     #[simd_test = "tbm"]
     #[cfg(not(target_arch = "x86"))]
     #[cfg_attr(rustfmt, rustfmt_skip)]
-    unsafe fn _blsic_u64() {
+    unsafe fn test_blsic_u64() {
         assert_eq!(
-            tbm::_blsic_u64(0b0101_0100u64),
+            _blsic_u64(0b0101_0100u64),
             0b1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1011u64
         );
         assert_eq!(
-            tbm::_blsic_u64(0u64),
+            _blsic_u64(0u64),
             0b1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111u64
         );
     }
 
     #[simd_test = "tbm"]
-    unsafe fn _t1mskc_u32() {
+    unsafe fn test_t1mskc_u32() {
         assert_eq!(
-            tbm::_t1mskc_u32(0b0101_0111u32),
+            _t1mskc_u32(0b0101_0111u32),
             0b1111_1111_1111_1111_1111_1111_1111_1000u32
         );
         assert_eq!(
-            tbm::_t1mskc_u32(0u32),
+            _t1mskc_u32(0u32),
             0b1111_1111_1111_1111_1111_1111_1111_1111u32
         );
     }
@@ -416,27 +416,27 @@ mod tests {
     #[simd_test = "tbm"]
     #[cfg(not(target_arch = "x86"))]
     #[cfg_attr(rustfmt, rustfmt_skip)]
-    unsafe fn _t1mksc_u64() {
+    unsafe fn test_t1mksc_u64() {
         assert_eq!(
-            tbm::_t1mskc_u64(0b0101_0111u64),
+            _t1mskc_u64(0b0101_0111u64),
             0b1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1000u64
         );
         assert_eq!(
-            tbm::_t1mskc_u64(0u64),
+            _t1mskc_u64(0u64),
             0b1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111_1111u64
         );
     }
 
     #[simd_test = "tbm"]
-    unsafe fn _tzmsk_u32() {
-        assert_eq!(tbm::_tzmsk_u32(0b0101_1000u32), 0b0000_0111u32);
-        assert_eq!(tbm::_tzmsk_u32(0b0101_1001u32), 0b0000_0000u32);
+    unsafe fn test_tzmsk_u32() {
+        assert_eq!(_tzmsk_u32(0b0101_1000u32), 0b0000_0111u32);
+        assert_eq!(_tzmsk_u32(0b0101_1001u32), 0b0000_0000u32);
     }
 
     #[simd_test = "tbm"]
     #[cfg(not(target_arch = "x86"))]
-    unsafe fn _tzmsk_u64() {
-        assert_eq!(tbm::_tzmsk_u64(0b0101_1000u64), 0b0000_0111u64);
-        assert_eq!(tbm::_tzmsk_u64(0b0101_1001u64), 0b0000_0000u64);
+    unsafe fn test_tzmsk_u64() {
+        assert_eq!(_tzmsk_u64(0b0101_1000u64), 0b0000_0111u64);
+        assert_eq!(_tzmsk_u64(0b0101_1001u64), 0b0000_0000u64);
     }
 }
