+++ coresimd/x86_64/abm.rs
@@ -26,15 +26,15 @@ use stdsimd_test::assert_instr;
 #[inline]
 #[target_feature(enable = "lzcnt")]
 #[cfg_attr(test, assert_instr(lzcnt))]
-pub unsafe fn _lzcnt_u32(x: u32) -> u32 {
-    x.leading_zeros()
+pub unsafe fn _lzcnt_u64(x: u64) -> u64 {
+    x.leading_zeros() as u64
 }
 
 /// Counts the bits that are set.
 #[inline]
 #[target_feature(enable = "popcnt")]
 #[cfg_attr(test, assert_instr(popcnt))]
-pub unsafe fn _popcnt32(x: i32) -> i32 {
+pub unsafe fn _popcnt64(x: i64) -> i32 {
     x.count_ones() as i32
 }
 
@@ -42,15 +42,15 @@ pub unsafe fn _popcnt32(x: i32) -> i32 {
 mod tests {
     use stdsimd_test::simd_test;
 
-    use coresimd::x86::i586::abm;
+    use coresimd::arch::x86_64::*;
 
     #[simd_test = "lzcnt"]
-    unsafe fn _lzcnt_u32() {
-        assert_eq!(abm::_lzcnt_u32(0b0101_1010), 25);
+    unsafe fn test_lzcnt_u64() {
+        assert_eq!(_lzcnt_u64(0b0101_1010), 57);
     }
 
     #[simd_test = "popcnt"]
-    unsafe fn _popcnt32() {
-        assert_eq!(abm::_popcnt32(0b0101_1010), 4);
+    unsafe fn test_popcnt64() {
+        assert_eq!(_popcnt64(0b0101_1010), 4);
     }
 }
