+++ coresimd/x86_64/bmi.rs
@@ -102,6 +102,7 @@ mod tests {
     use stdsimd_test::simd_test;
 
     use coresimd::x86::*;
+    use coresimd::x86_64::*;
 
     #[simd_test = "bmi"]
     unsafe fn test_bextr_u64() {
