+++ coresimd/x86_64/sse42.rs
@@ -20,7 +20,7 @@ pub unsafe fn _mm_crc32_u64(crc: u64, v: u64) -> u64 {
 
 #[cfg(test)]
 mod tests {
-    use coresimd::x86::*;
+    use coresimd::arch::x86_64::*;
 
     use stdsimd_test::simd_test;
